refactor: Redux logic and provider moved to index.jsx
